Refactor admin layout to use AdminGuard component.
Introduced AdminGuard to handle admin access control logic, reducing redundancy in layout. Added metadata management for the admin section, improving structure and maintainability. Streamlined the admin page setup for better readability and scalability.
